requirements.txt
Django==2.0
djangorestframework==3.7.7
Markdown==2.6.11
django-filter==1.1.0
django-allauth==0.35.0
